+++ voxygen/anim/src/character/repeater.rs
@@ -10,7 +10,7 @@ impl Animation for RepeaterAnimation {
         Option<ToolKind>,
         Option<ToolKind>,
         Vec3<f32>,
-        f64,
+        f32,
         Option<StageSection>,
     );
     type Skeleton = CharacterSkeleton;
@@ -23,7 +23,7 @@ impl Animation for RepeaterAnimation {
     fn update_skeleton_inner(
         skeleton: &Self::Skeleton,
         (active_tool_kind, _second_tool_kind, _velocity, _global_time, stage_section): Self::Dependency,
-        anim_time: f64,
+        anim_time: f32,
         rate: &mut f32,
         s_a: &SkeletonAttr,
     ) -> Self::Skeleton {
@@ -31,10 +31,10 @@ impl Animation for RepeaterAnimation {
         let mut next = (*skeleton).clone();
 
         let (move1, move2, move3, _move4) = match stage_section {
-            Some(StageSection::Movement) => (anim_time as f32, 0.0, 0.0, 0.0),
-            Some(StageSection::Buildup) => (1.0, anim_time as f32, 0.0, 0.0),
-            Some(StageSection::Shoot) => (1.0, 1.0, anim_time as f32, 0.0),
-            Some(StageSection::Recover) => (1.1, 1.0, 1.0, anim_time as f32),
+            Some(StageSection::Movement) => (anim_time, 0.0, 0.0, 0.0),
+            Some(StageSection::Buildup) => (1.0, anim_time, 0.0, 0.0),
+            Some(StageSection::Shoot) => (1.0, 1.0, anim_time, 0.0),
+            Some(StageSection::Recover) => (1.1, 1.0, 1.0, anim_time),
             _ => (0.0, 0.0, 0.0, 0.0),
         };
