+++ voxygen/anim/src/quadruped_low/shoot.rs
@@ -7,7 +7,7 @@ use common::states::utils::StageSection;
 pub struct ShootAnimation;
 
 impl Animation for ShootAnimation {
-    type Dependency = (f32, f64, Option<StageSection>);
+    type Dependency = (f32, f32, Option<StageSection>);
     type Skeleton = QuadrupedLowSkeleton;
 
     #[cfg(feature = "use-dyn-lib")]
@@ -17,16 +17,16 @@ impl Animation for ShootAnimation {
     fn update_skeleton_inner(
         skeleton: &Self::Skeleton,
         (velocity, _global_time, stage_section): Self::Dependency,
-        anim_time: f64,
+        anim_time: f32,
         _rate: &mut f32,
         s_a: &SkeletonAttr,
     ) -> Self::Skeleton {
         let mut next = (*skeleton).clone();
 
         let (movement1, movement2, movement3) = match stage_section {
-            Some(StageSection::Buildup) => (anim_time as f32, 0.0, 0.0),
-            Some(StageSection::Swing) => (1.0, (anim_time as f32).powf(0.25), 0.0),
-            Some(StageSection::Recover) => (1.0, 1.0, anim_time as f32),
+            Some(StageSection::Buildup) => (anim_time, 0.0, 0.0),
+            Some(StageSection::Swing) => (1.0, anim_time.powf(0.25), 0.0),
+            Some(StageSection::Recover) => (1.0, 1.0, anim_time),
             _ => (0.0, 0.0, 0.0),
         };
